// frontend/src/app/core/interfaces/products.interfaces.ts
// Generated by https://quicktype.io

export interface FormProducts {
  page: number;
  country?: string;
  category?: string;
}

// Generated by https://quicktype.io

export interface ProductsResponse {
  id: string;
  name: string;
  description: string;
  stock: number;
  image: string;
  price: number;
  weight: number;
  country: string;
  minStock: number;
  state: boolean;
  category: Category;
  quantity?: number;
}

export interface Category {
  id: string;
  name: string;
  description: string;
  number: string;
  state: boolean;
}

export interface CartProduct {
  id: string;
  name: string;
  image: string;
  price: number;
  stock: number;
  quantity: number;
  weight: number;
}

// Generated by https://quicktype.io

export interface ProductById {
  id: string;
  name: string;
  description: string;
  stock: number;
  image: string;
  price: number;
  weight: number;
  country: string;
  minStock: number;
  state: string;
  category: Category;
}

export interface Category {
  id: string;
  name: string;
  description: string;
  number: string;
  state: boolean;
}